// src/generated/models/Sd_goodsreceiptsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Sd_goodsreceiptssd_kind = {
  100000000: 'FullReceipt',
  100000001: 'PartialReceipt',
  100000002: 'MilestoneConfirmed',
  100000003: 'Rejected'
} as const;
export type Sd_goodsreceiptssd_kind = keyof typeof Sd_goodsreceiptssd_kind;
export const Sd_goodsreceiptsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Sd_goodsreceiptsstatecode = keyof typeof Sd_goodsreceiptsstatecode;
export const Sd_goodsreceiptsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Sd_goodsreceiptsstatuscode = keyof typeof Sd_goodsreceiptsstatuscode;

export interface Sd_goodsreceiptsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  sd_company?: string;
  sd_confirmedbyname?: string;
  sd_confirmedbypersonnelnumber?: string;
  sd_goodsreceiptid: string;
  sd_kind?: Sd_goodsreceiptssd_kind;
  sd_name?: string;
  sd_notes?: string;
  sd_purchaseorderline?: number;
  sd_purchaseordernumber: string;
  sd_receiveddate?: string;
  sd_receivedquantity?: number;
  sd_requisitionnumber?: string;
  sd_vendoraccount?: string;
  statecode: Sd_goodsreceiptsstatecode;
  statuscode?: Sd_goodsreceiptsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Sd_goodsreceipts extends Sd_goodsreceiptsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  sd_attachment?: string;
  sd_attachment_name?: string;
  sd_kindname?: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Sd_goodsreceiptsUploadColumnName = 'sd_attachment';